+++ app/src/main/AndroidManifest.xml
@@ -14,7 +14,7 @@
         android:theme="@style/AppTheme"
         tools:ignore="GoogleAppIndexingWarning">
         <activity android:name=".ads.activity.AdsDetailActivity"></activity>
-        <activity android:name=".mypage.CashActivity" /> <!-- 광고 -->
+        <activity android:name=".mypage.cash.CashActivity" /> <!-- 광고 -->
         <activity android:name=".ads.category.CategoryActivity" /> <!-- 메인 -->
         <activity android:name=".main.MainActivity">
             <intent-filter>
