--- app/src/main/java/org/techtown/crecker/mypage/UsageRecord.kt
@@ -1,10 +1,0 @@
-package org.techtown.crecker.mypage
-
-import java.util.*
-
-data class UsageRecord(
-    var title: String = "",
-    var io: String = "",
-    var money: Int = 0,
-    var date: Date = Date()
-)
