+++ app/src/main/java/org/techtown/crecker/mypage/cash/InfoDialog.kt
@@ -1,0 +1,1 @@
+package org.techtown.crecker.mypage.cash
