+++ app/src/main/java/org/techtown/crecker/mypage/contents/usage/UsageRecord.kt
@@ -1,0 +1,10 @@
+package org.techtown.crecker.mypage.contents.usage
+
+import java.util.*
+
+data class UsageRecord(
+    var title: String = "",
+    var io: String = "",
+    var money: Int = 0,
+    var date: Date = Date()
+)
